+++ vacations/src/main/resources/templates/vacations.html
@@ -113,7 +113,7 @@
 
         </data-table>
 
-        <year-calendar :year="2020" ></year-calendar>
+        <year-calendar :year="2020"  :events="calendarData"></year-calendar>
 
 
         <tmodal-confirm
