+++ webapp/src/main/resources/public/js/calendar.js
@@ -38,16 +38,16 @@ Vue.component('calendar', {
             <tr>
               <td style="width: 10rem; white-space: nowrap;" rowspan="2" v-if="showColName === true" > {{ name }}</td>
               <td v-for="day in daysInMonth" v-bind:style="[
-               (day.date.getDay() === 0 || day.date.getDay() === 6) ? { 'background-color' : 'lightgrey' } : {}, 
-               (day.morningEvent !== undefined && selectColor(day.morningEvent)) ? {'background-color' : selectColor(day.morningEvent) } : {}
+               (day.morningEvent !== undefined && selectColor(day.morningEvent)) ? {'background-color' : selectColor(day.morningEvent) } : {},
+               (day.date.getDay() === 0 || day.date.getDay() === 6) ? { 'background-color' : 'lightgrey' } : {} 
                ] " v-bind:data-label="day.date.toDateString()" ></td>  
               
                <td v-for="index in (31 - daysInMonth.length)" :key="index" style="background-color: rgba(0,0,0,.05)" rowspan="2"></td>        
             </tr>
             <tr>
               <td v-for="day in daysInMonth" v-bind:style="[
-               (day.date.getDay() === 0 || day.date.getDay() === 6) ? { 'background-color' : 'lightgrey' } : {}, 
-               (day.afternoonEvent !== undefined && selectColor(day.afternoonEvent)) ? {'background-color' : selectColor(day.afternoonEvent) } : {}
+               (day.afternoonEvent !== undefined && selectColor(day.afternoonEvent)) ? {'background-color' : selectColor(day.afternoonEvent) } : {},
+               (day.date.getDay() === 0 || day.date.getDay() === 6) ? { 'background-color' : 'lightgrey' } : {} 
                ] " v-bind:data-label="day.date.toDateString()" ></td>        
             </tr>
         </table>
@@ -82,11 +82,21 @@ Vue.component('calendar', {
 
 Vue.component('year-calendar', {
     props: {
-        year : Number
+        year : Number,
+        events :  {
+            type : Array,
+            default : () => ([
+                {
+                    date : new Date().toDateString().substr(0,10),
+                    value : 1,
+                    type : 1
+                }
+            ])
+        }
     },
     template: `
         <div>
-            <calendar v-for="month in monthsInYear" :name="monthNames[month]" :show-header="month === 0" :year="year" :month="month"></calendar>
+            <calendar v-for="month in monthsInYear" :name="monthNames[month]" :show-header="month === 0" :year="year" :month="month" :events="events"></calendar>
         </div>
        `,
     computed: {
